backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
anthropic==0.28.0
reportlab==4.2.2
streamlit==1.35.0
pyyaml==6.0.1
python-dotenv==1.0.1
apscheduler==3.10.4
pytest==8.2.0
pytest-asyncio==0.23.7
